Bug: PRD Reviewer: Peter Comment: Version controls - Set file upload size per-user and per-group (UI) only

+++ ssf/main/src/org/kablink/teaming/jobs/DefaultFileVersionAgingDelete.java
@@ -53,7 +53,7 @@ public class DefaultFileVersionAgingDelete extends SSCronTriggerJob
 	 
     public void doExecute(JobExecutionContext context) throws JobExecutionException {
     	Long zoneId = RequestContextHolder.getRequestContext().getZoneId();
-    	Integer fileVersionsMaxAge = coreDao.loadZoneConfig(zoneId).getFileVersionsMaxAge();
+    	Long fileVersionsMaxAge = coreDao.loadZoneConfig(zoneId).getFileVersionsMaxAge();
 		if (fileVersionsMaxAge == null) {
 			logger.debug("File version aging is not enabled for zone " + RequestContextHolder.getRequestContext().getZoneName());
 	   		context.put(CleanupJobListener.CLEANUPSTATUS, CleanupJobListener.DeleteJob);
